Fix null pointer dereference in RegsArm.
Fix RegsArm::GetPcAdjustment to check for an invalid elf before trying
to read memory.
Modify the tests for this so it crashes without this change.

Also modify the GetPcAdjustment for all different architectures so
that unless the relative pc is too small, it will return the minimum
amount that should be adjusted. This is to handle cases where we still
want to adjust the pc but it's in an invalid elf. Mostly this is for
handling cases when the pc is in jit gdb debug code so that we use the
right unwind information.

Bug: 77233204

Test: Passes unit tests for libbacktrace/libunwindstack.
Change-Id: Id73609adaf3b80a583584441de228156fec3afa7
(cherry picked from commit a62f988495ed209909e3e7a27758effdae684f96 [formerly 6dbc28ece3ab7cadd0087b4dc31ba9a2986545f0])

Former-commit-id: 72f2376d8d8f7f0ce03a484038c85db5ede1f09c [formerly fd6b7284ee4154bdab11da7d332920dc8b6aa342]
Former-commit-id: 939163b3c06db9ef804bd8aab9477491a69266e3